+++ src/main/scala/com/bharatsim/engine/graph/ingestion/NodeExpander.scala
@@ -7,12 +7,12 @@ import com.bharatsim.engine.utils.Utils.fetchClassName
 import scala.reflect.ClassTag
 
 class NodeExpander {
-  def expand[T <: Node : ClassTag](nodeRef: Int, node: T): GraphData = {
+  def expand[T <: Node : ClassTag](nodeRef: Long, node: T): GraphData = {
     val label = fetchClassName[T]
     expand(label, nodeRef, node)
   }
 
-  def expand[T <: Node](label: String, nodeRef: Int, node: T): GraphData = {
+  def expand[T <: Node](label: String, nodeRef: Long, node: T): GraphData = {
     node match {
       case statefulAgent: StatefulAgent =>
