+++ src/main/scala/com/bharatsim/engine/models/Identity.scala
@@ -1,5 +1,5 @@
 package com.bharatsim.engine.models
 
 trait Identity {
-  var internalId: Int
+  var internalId: Long
 }
